move files and rename
