+++ WalletWasabi.Gui/Controls/WalletExplorer/BuildTabView.xaml.cs
@@ -1,0 +1,19 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace WalletWasabi.Gui.Controls.WalletExplorer
+{
+	public class BuildTabView : UserControl
+	{
+		public BuildTabView()
+		{
+			this.InitializeComponent();
+		}
+
+		private void InitializeComponent()
+		{
+			AvaloniaXamlLoader.Load(this);
+		}
+	}
+}
