+++ WalletWasabi.Gui/Controls/WalletExplorer/SendTabView.xaml
@@ -1,6 +1,6 @@
 ﻿<UserControl xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:controls="clr-namespace:WalletWasabi.Gui.Controls;assembly=WalletWasabi.Gui"
+             xmlns:walletcontrols="clr-namespace:WalletWasabi.Gui.Controls.WalletExplorer;assembly=WalletWasabi.Gui"
              x:Class="WalletWasabi.Gui.Controls.WalletExplorer.SendTabView">
-  <controls:SendTabBase />
+  <walletcontrols:SendTabBaseView />
 </UserControl>
